requirements.txt
crewai
crewai-tools
python-dotenv
mistralai
vnstock
pandas
numpy
langchain-google-genai 